requirements.txt
boto3
groq
pdfminer.six
python-dotenv